requirements.txt
pandas==2.2.2
pyarrow==17.0.0
numpy==1.26.4
matplotlib==3.9.0
scikit-learn==1.5.1
jupyterlab==4.2.3
seaborn==0.13.2
plotly==5.22.0
openpyxl==3.1.2
google-cloud-bigquery==3.38.0
